+++ backend/src/main/java/com/woowacourse/moamoa/study/domain/study/Study.java
@@ -1,9 +1,9 @@
-package com.woowacourse.moamoa.study.domain;
+package com.woowacourse.moamoa.study.domain.study;
 
 import static javax.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
-import com.woowacourse.moamoa.study.domain.filter.StudyFilter;
+import com.woowacourse.moamoa.study.domain.studyfilter.StudyFilter;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
