+++ backend/src/main/java/com/woowacourse/moamoa/study/domain/study/repository/StudyRepository.java
@@ -1,6 +1,6 @@
-package com.woowacourse.moamoa.study.domain.repository;
+package com.woowacourse.moamoa.study.domain.study.repository;
 
-import com.woowacourse.moamoa.study.domain.Study;
+import com.woowacourse.moamoa.study.domain.study.Study;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
