+++ backend/src/main/java/com/woowacourse/moamoa/study/domain/studyfilter/repository/StudyFilterRepository.java
@@ -1,8 +1,8 @@
-package com.woowacourse.moamoa.study.domain.filter.repository;
+package com.woowacourse.moamoa.study.domain.studyfilter.repository;
 
-import com.woowacourse.moamoa.study.domain.filter.StudyFilter;
-import com.woowacourse.moamoa.study.domain.filter.StudySearchCondition;
-import com.woowacourse.moamoa.study.domain.filter.StudySlice;
+import com.woowacourse.moamoa.study.domain.studyfilter.StudyFilter;
+import com.woowacourse.moamoa.study.domain.studyfilter.StudySearchCondition;
+import com.woowacourse.moamoa.study.domain.studyfilter.StudySlice;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
