+++ backend/src/main/java/com/woowacourse/moamoa/study/service/StudyFilterService.java
@@ -5,13 +5,13 @@ import static java.util.stream.Collectors.toList;
 import com.woowacourse.moamoa.filter.domain.Filter;
 import com.woowacourse.moamoa.filter.domain.repository.FilterRepository;
 import com.woowacourse.moamoa.study.controller.request.FilterRequest;
-import com.woowacourse.moamoa.study.domain.Study;
-import com.woowacourse.moamoa.study.domain.repository.StudyRepository;
+import com.woowacourse.moamoa.study.domain.study.Study;
+import com.woowacourse.moamoa.study.domain.study.repository.StudyRepository;
 import com.woowacourse.moamoa.study.service.response.StudiesResponse;
 import com.woowacourse.moamoa.study.service.response.StudyResponse;
-import com.woowacourse.moamoa.study.domain.filter.StudySearchCondition;
-import com.woowacourse.moamoa.study.domain.filter.StudySlice;
-import com.woowacourse.moamoa.study.domain.filter.repository.StudyFilterRepository;
+import com.woowacourse.moamoa.study.domain.studyfilter.StudySearchCondition;
+import com.woowacourse.moamoa.study.domain.studyfilter.StudySlice;
+import com.woowacourse.moamoa.study.domain.studyfilter.repository.StudyFilterRepository;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
