+++ backend/src/main/java/com/woowacourse/moamoa/study/service/StudyService.java
@@ -1,7 +1,7 @@
 package com.woowacourse.moamoa.study.service;
 
-import com.woowacourse.moamoa.study.domain.Study;
-import com.woowacourse.moamoa.study.domain.repository.StudyRepository;
+import com.woowacourse.moamoa.study.domain.study.Study;
+import com.woowacourse.moamoa.study.domain.study.repository.StudyRepository;
 import com.woowacourse.moamoa.study.service.response.StudiesResponse;
 import com.woowacourse.moamoa.study.service.response.StudyResponse;
 import java.util.List;
