+++ backend/src/test/java/com/woowacourse/moamoa/study/domain/study/repository/StudyRepositoryTest.java
@@ -1,10 +1,10 @@
-package com.woowacourse.moamoa.study.repository;
+package com.woowacourse.moamoa.study.domain.study.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
 
-import com.woowacourse.moamoa.study.domain.Study;
-import com.woowacourse.moamoa.study.domain.repository.StudyRepository;
+import com.woowacourse.moamoa.study.domain.study.Study;
+import com.woowacourse.moamoa.study.domain.study.repository.StudyRepository;
 import java.util.List;
 import java.util.stream.Stream;
 import org.assertj.core.groups.Tuple;
